Melee Weapon Fix
Partial class separation to handle attack logic and damage dealing logic separately.
Fix for observable RPC. Now the owner is excluded to prevent a double call, as it is being passed through the server.
A list was added to keep track of objects that have already taken damage, to avoid dealing damage to them again with the same hit when entering and exiting the trigger.